backend/requirements.txt
ultralytics>=8.3.0
opencv-python>=4.10.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
numpy>=2.0.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
pydantic>=2.0.0
httpx
python-dotenv
